requirements.txt
matplotlib==3.10.7
numpy==1.23.3
openai==2.3.0
pandas==1.5.0
plotly==5.22.0
python-dotenv==1.1.1
scikit_learn==1.7.2
seaborn==0.13.2
streamlit==1.28.2
xgboost==3.0.5
